# frontend/src/public/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "RIS API",
    "contact" : {
      "name" : "DigitalService GmbH des Bundes",
      "url" : "https://digitalservice.bund.de"
    },
    "version" : "1"
  },
  "servers" : [ {
    "url" : "https://testphase.rechtsinformationen.bund.de",
    "description" : "Public API (preview)"
  } ],
...
  "paths" : {
...
    "/v1/literature" : {
      "get" : {
        "tags" : [ "Literature" ],
        "summary" : "List and search literature",
        "description" : "The endpoint returns a list of literature from our database. The list is paginated and can be filtered and sorted.",
        "operationId" : "searchLiterature",
        "parameters" : [ {
          "name" : "documentNumber",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "yearOfPublication",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "documentType",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "author",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "collaborator",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "searchTerm",
          "in" : "query",
          "description" : "Searches for the given tokens in searchTerm. If searchTerm contains more than one token, all tokens must be in the document for the document to match.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "Searches for the given tokens in searchTerm. If searchTerm contains more than one token, all tokens must be in the document for the document to match."
          }
        }, {
          "name" : "dateFrom",
          "in" : "query",
          "description" : "The from (greater than or equal) parameter returns all entities where date is later than, or equal to, the given date.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date",
            "description" : "The from (greater than or equal) parameter returns all entities where date is later than, or equal to, the given date."
          }
        }, {
          "name" : "dateTo",
          "in" : "query",
          "description" : "The to (less than or equal) parameter returns all entities where date is earlier than, or equal to, the given date.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date",
            "description" : "The to (less than or equal) parameter returns all entities where date is earlier than, or equal to, the given date."
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The number of entities per page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 300,
            "minimum" : 1
          },
          "example" : 100
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "The number of the page to request. The page starts with the value 0",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "example" : 0
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date and documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date and documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
            "pattern" : "^-?(|default|date|DATUM|documentNumber)$"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionSchemaSearchMemberSchemaLiteratureSearchSchema"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
    "/v1/document/lucene-search/literature" : {
      "get" : {
        "tags" : [ "Literature" ],
        "summary" : "Advanced search using Lucene query syntax",
        "operationId" : "literatureSearch",
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "description" : "The query filter based on Lucene query syntax",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The number of entities per page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 300,
            "minimum" : 1
          },
          "example" : 100
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "The number of the page to request. The page starts with the value 0",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "example" : 0
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date and documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date and documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
            "pattern" : "^-?(|default|date|DATUM|documentNumber)$"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionSchemaSearchMemberSchemaLiteratureSearchSchema"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/v1/document/lucene-search/legislation" : {
      "get" : {
        "tags" : [ "Legislation" ],
        "summary" : "Advanced search using Lucene query syntax",
        "operationId" : "searchLegislation",
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "description" : "The query filter based on Lucene query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The number of entities per page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 300,
            "minimum" : 1
          },
          "example" : 100
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "The number of the page to request. The page starts with the value 0",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "example" : 0
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date, temporalCoverageFrom, legislationIdentifier and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date, temporalCoverageFrom, legislationIdentifier and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
            "pattern" : "^-?(|default|date|temporalCoverageFrom|legislationIdentifier|DATUM)$"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionSchemaSearchMemberSchemaLegislationExpressionSearchSchema"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/v1/document/lucene-search/case-law" : {
      "get" : {
        "tags" : [ "Case Law" ],
        "summary" : "Advanced search using Lucene query syntax",
        "operationId" : "caseLawSearch",
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "description" : "The query filter based on Lucene query syntax",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The number of entities per page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 300,
            "minimum" : 1
          },
          "example" : 100
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "The number of the page to request. The page starts with the value 0",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "example" : 0
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date, courtName, documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date, courtName, documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
            "pattern" : "^-?(|default|date|DATUM|courtName|documentNumber)$"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionSchemaSearchMemberSchemaCaseLawSearchSchema"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/v1/document/lucene-search/administrative-directive" : {
      "get" : {
        "tags" : [ "AdministrativeDirective" ],
        "summary" : "Advanced search using Lucene query syntax",
        "operationId" : "literatureSearch_1",
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "description" : "The query filter based on Lucene query syntax",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The number of entities per page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 300,
            "minimum" : 1
          },
          "example" : 100
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "The number of the page to request. The page starts with the value 0",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "example" : 0
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date and documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date and documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
            "pattern" : "^-?(|default|date|DATUM|documentNumber)$"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionSchemaSearchMemberSchemaAdministrativeDirectiveSearchSchema"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
    "/v1/administrative-directive" : {
      "get" : {
        "tags" : [ "AdministrativeDirective" ],
        "summary" : "List and search administrative directives",
        "description" : "The endpoint returns a list of administrative directives from our database. The list is paginated and can be filtered and sorted.",
        "operationId" : "searchAdministrativeDirective",
        "parameters" : [ {
          "name" : "documentNumber",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "searchTerm",
          "in" : "query",
          "description" : "Searches for the given tokens in searchTerm. If searchTerm contains more than one token, all tokens must be in the document for the document to match.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "Searches for the given tokens in searchTerm. If searchTerm contains more than one token, all tokens must be in the document for the document to match."
          }
        }, {
          "name" : "dateFrom",
          "in" : "query",
          "description" : "The from (greater than or equal) parameter returns all entities where date is later than, or equal to, the given date.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date",
            "description" : "The from (greater than or equal) parameter returns all entities where date is later than, or equal to, the given date."
          }
        }, {
          "name" : "dateTo",
          "in" : "query",
          "description" : "The to (less than or equal) parameter returns all entities where date is earlier than, or equal to, the given date.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date",
            "description" : "The to (less than or equal) parameter returns all entities where date is earlier than, or equal to, the given date."
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The number of entities per page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 300,
            "minimum" : 1
          },
          "example" : 100
        }, {
          "name" : "pageIndex",
          "in" : "query",
          "description" : "The number of the page to request. The page starts with the value 0",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "example" : 0
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date and documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "The field to sort the results by. Default is the relevance score calculated by OpenSearch. Valid usage of the sort field are : date and documentNumber and not setting the sort field (sort by relevance descending).Add a leading - to set the order to descending (-date)",
            "pattern" : "^-?(|default|date|DATUM|documentNumber)$"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CollectionSchemaSearchMemberSchemaAdministrativeDirectiveSearchSchema"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CollectionSchemaSearchMemberSchemaCaseLawSearchSchema" : {
        "type" : "object",
        "properties" : {
          "@type" : {
            "type" : "string",
            "example" : "hydra:Collection"
          },
          "@context" : {
            "type" : "string"
          },
          "@id" : {
            "type" : "string",
            "example" : "/v1/document?pageIndex=0&size=5"
          },
          "totalItems" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 1
          },
          "member" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SearchMemberSchemaCaseLawSearchSchema"
            }
          },
          "view" : {
            "$ref" : "#/components/schemas/PartialCollectionViewSchema"
          }
        },
        "required" : [ "@context", "@id", "member", "totalItems", "view" ]
      },
...
      "CollectionSchemaSearchMemberSchemaLiteratureSearchSchema" : {
        "type" : "object",
        "properties" : {
          "@type" : {
            "type" : "string",
            "example" : "hydra:Collection"
          },
          "@context" : {
            "type" : "string"
          },
          "@id" : {
            "type" : "string",
            "example" : "/v1/document?pageIndex=0&size=5"
          },
          "totalItems" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 1
          },
          "member" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SearchMemberSchemaLiteratureSearchSchema"
            }
          },
          "view" : {
            "$ref" : "#/components/schemas/PartialCollectionViewSchema"
          }
        },
        "required" : [ "@context", "@id", "member", "totalItems", "view" ]
      },
...
      "CollectionSchemaSearchMemberSchemaLegislationExpressionSearchSchema" : {
        "type" : "object",
        "properties" : {
          "@type" : {
            "type" : "string",
            "example" : "hydra:Collection"
          },
          "@context" : {
            "type" : "string"
          },
          "@id" : {
            "type" : "string",
            "example" : "/v1/document?pageIndex=0&size=5"
          },
          "totalItems" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 1
          },
          "member" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SearchMemberSchemaLegislationExpressionSearchSchema"
            }
          },
          "view" : {
            "$ref" : "#/components/schemas/PartialCollectionViewSchema"
          }
        },
        "required" : [ "@context", "@id", "member", "totalItems", "view" ]
      },
...
    }
  }
}